Added iActions interface for processing transactions
Improved abstraction in Bank class by creating abstract methods for
prompting the user for transactions and getting the user's input and
processing the transaction. Account class gives a concrete
implementation of getTransactionValue and the individual account
subclasses provide a concrete implementation of the processTransaction
method that is unique to account type.